src/main/java/me/sup2is/board/board/BoardController.java
package me.sup2is.board.board;

import javax.validation.Valid;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseBody;

import me.sup2is.board.model.BoardForm;

@Controller
@RequestMapping("/board")
public class BoardController {
	
	@Autowired
	private BoardService boardService;

	@GetMapping("/write")
	public String writePage() {
		return "write";
	}
	
	@PostMapping(value = "/write")
	@ResponseBody
	public ResponseEntity<String> write(@RequestBody @Valid BoardForm boardForm) {
		try {
			boardService.write(boardForm.getBoardEntity());
			return new ResponseEntity<>("success", HttpStatus.OK);
		} catch (Exception e) {
			return new ResponseEntity<>("error", HttpStatus.INTERNAL_SERVER_ERROR);
		}
	}
	
}
